create controller for resource route
